Delete modified rules files + merge match command and match strategy
