+++ StaticAnalyzer/views/android/rules/android_apis.py
@@ -37,393 +37,310 @@ Rule Format.
 
 """
 
+from StaticAnalyzer.views.matchers import (
+    RegexAnd,
+    SingleRegex,
+    SingleString,
+    StringAnd,
+    StringAndOr,
+)
 from StaticAnalyzer.views.rules_properties import (
     InputCase,
-    Match,
-    MatchType,
 )
-
 APIS = [
     {
         'desc': 'Loading Native Code (Shared Library) ',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'System\.loadLibrary\(|System\.load\(',
+        'type': SingleRegex.__name__,
+        'match': r'System\.loadLibrary\(|System\.load\(',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get System Service',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'getSystemService',
+        'type': SingleRegex.__name__,
+        'match': r'getSystemService',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Dynamic Class and Dexloading',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'dalvik\.system\.DexClassLoader|'
-                   r'java\.security\.ClassLoader|'
-                   r'java\.net\.URLClassLoader|'
-                   r'java\.security\.SecureClassLoader'),
+        'type': SingleRegex.__name__,
+        'match': (r'dalvik\.system\.DexClassLoader|'
+                  r'java\.security\.ClassLoader|'
+                  r'java\.net\.URLClassLoader|'
+                  r'java\.security\.SecureClassLoader'),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Java Reflection',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'java\.lang\.reflect\.Method|'
-                   r'java\.lang\.reflect\.Field|Class\.forName'),
+        'type': SingleRegex.__name__,
+        'match': (r'java\.lang\.reflect\.Method|'
+                  r'java\.lang\.reflect\.Field|Class\.forName'),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Crypto',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'javax\.crypto|kalium\.crypto|bouncycastle\.crypto',
+        'type': SingleRegex.__name__,
+        'match': r'javax\.crypto|kalium\.crypto|bouncycastle\.crypto',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Starting Activity',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'startActivity\(|startActivityForResult\(',
+        'type': SingleRegex.__name__,
+        'match': r'startActivity\(|startActivityForResult\(',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Starting Service',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'startService\(|bindService\(',
+        'type': SingleRegex.__name__,
+        'match': r'startService\(|bindService\(',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Sending Broadcast',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'sendBroadcast\(|'
-                   r'sendOrderedBroadcast\(|sendStickyBroadcast\('),
+        'type': SingleRegex.__name__,
+        'match': (r'sendBroadcast\(|'
+                  r'sendOrderedBroadcast\(|sendStickyBroadcast\('),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Local File I/O Operations',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'OpenFileOutput|getSharedPreferences|'
-                   r'SharedPreferences\.Editor|getCacheDir|'
-                   r'getExternalStorageState|openOrCreateDatabase'),
+        'type': SingleRegex.__name__,
+        'match': (r'OpenFileOutput|getSharedPreferences|'
+                  r'SharedPreferences\.Editor|getCacheDir|'
+                  r'getExternalStorageState|openOrCreateDatabase'),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Inter Process Communication',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'IRemoteService|IRemoteService\.Stub|IBinder|Intent',
+        'type': SingleRegex.__name__,
+        'match': r'IRemoteService|IRemoteService\.Stub|IBinder|Intent',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'HTTP Requests, Connections and Sessions',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'http\.client\.HttpClient|net\.http\.AndroidHttpClient|'
-                   r'http\.impl\.client\.AbstractHttpClient'),
+        'type': SingleRegex.__name__,
+        'match': (r'http\.client\.HttpClient|net\.http\.AndroidHttpClient|'
+                  r'http\.impl\.client\.AbstractHttpClient'),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'HTTP Connection',
-        'type': MatchType.regex,
-        'match': Match.regex_and,
-        'regex1': r'HttpURLConnection|org\.apache\.http',
-        'regex2': r'openConnection|connect|HttpRequest',
+        'type': RegexAnd.__name__,
+        'match': [r'HttpURLConnection|org\.apache\.http',
+                  r'openConnection|connect|HttpRequest'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Load and Manipulate Dex Files',
-        'type': MatchType.regex,
-        'match': Match.regex_and,
-        'regex1': (r'dalvik\.system\.PathClassLoader|'
+        'type': RegexAnd.__name__,
+        'match': [(r'dalvik\.system\.PathClassLoader|'
                    r'dalvik\.system\.DexFile|dalvik\.system\.DexPathList'),
-        'regex2': r'loadDex|loadClass|DexClassLoader|loadDexFile',
+                  r'loadDex|loadClass|DexClassLoader|loadDexFile'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Content Provider',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'android.content.ContentProvider',
+        'type': SingleString.__name__,
+        'match': 'android.content.ContentProvider',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Obfuscation',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'utils.AESObfuscator',
-        'string2': 'getObfuscator',
+        'type': StringAnd.__name__,
+        'match': ['utils.AESObfuscator', 'getObfuscator'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Execute OS Command',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'getRuntime().exec(',
-        'string2': 'getRuntime(',
+        'type': StringAnd.__name__,
+        'match': ['getRuntime().exec(', 'getRuntime('],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Android Keystore',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'security.KeyStore',
-        'string2': 'Keystore.getInstance(',
+        'type': StringAnd.__name__,
+        'match': ['security.KeyStore', 'Keystore.getInstance('],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'TCP Server Socket',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'ServerSocket',
-        'string2': 'net.ServerSocket',
+        'type': StringAnd.__name__,
+        'match': ['ServerSocket', 'net.ServerSocket'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'TCP Socket',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'Socket',
-        'string2': 'net.Socket',
+        'type': StringAnd.__name__,
+        'match': ['Socket', 'net.Socket'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'UDP Datagram Packet',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'DatagramPacket',
-        'string2': 'net.DatagramPacket',
+        'type': StringAnd.__name__,
+        'match': ['DatagramPacket', 'net.DatagramPacket'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'UDP Datagram Socket',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'DatagramSocket',
-        'string2': 'net.DatagramSocket',
+        'type': StringAnd.__name__,
+        'match': ['DatagramSocket', 'net.DatagramSocket'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView JavaScript Interface',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'addJavascriptInterface',
-        'string2': 'WebView',
-        'string3': 'android.webkit',
+        'type': StringAnd.__name__,
+        'match': ['addJavascriptInterface', 'WebView', 'android.webkit'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView GET Request',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'WebView',
-        'string2': 'loadData',
-        'string3': 'android.webkit',
+        'type': StringAnd.__name__,
+        'match': ['WebView', 'loadData', 'android.webkit'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView POST Request',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'WebView',
-        'string2': 'postUrl',
-        'string3': 'android.webkit',
+        'type': StringAnd.__name__,
+        'match': ['WebView', 'postUrl', 'android.webkit'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Android Notifications',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'app.NotificationManager',
-        'string2': 'notify',
+        'type': StringAnd.__name__,
+        'match': ['app.NotificationManager', 'notify'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get Cell Information',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getAllCellInfo',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getAllCellInfo'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get Cell Location',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getCellLocation',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getCellLocation'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get Subscriber ID',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getSubscriberId',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getSubscriberId'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get Device ID, IMEI,MEID/ESN etc.',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getDeviceId',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getDeviceId'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get Software Version, IMEI/SV etc.',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getDeviceSoftwareVersion',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getDeviceSoftwareVersion'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get SIM Serial Number',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getSimSerialNumber',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getSimSerialNumber'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get SIM Provider Details',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getSimOperator',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getSimOperator'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Get SIM Operator Name',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'telephony.TelephonyManager',
-        'string2': 'getSimOperatorName',
+        'type': StringAnd.__name__,
+        'match': ['telephony.TelephonyManager', 'getSimOperatorName'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Query Database of SMS, Contacts etc.',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'content.ContentResolver',
-        'string2': 'query',
+        'type': StringAnd.__name__,
+        'match': ['content.ContentResolver', 'query'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Query Database of SMS, Contacts etc.',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'content.ContentResolver',
-        'string2': 'query',
+        'type': StringAnd.__name__,
+        'match': ['content.ContentResolver', 'query'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Base64 Decode',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'android.util.Base64',
-        'string2': '.decode',
+        'type': StringAnd.__name__,
+        'match': ['android.util.Base64', '.decode'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Send SMS',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'telephony.SmsManager',
-        'string_or1': 'sendMultipartTextMessage',
-        'string_or2': 'sendTextMessage',
-        'string_or3': 'vnd.android-dir/mms-sms',
+        'type': StringAndOr.__name__,
+        'match': ['telephony.SmsManager',
+                  ['sendMultipartTextMessage',
+                   'sendTextMessage', 'vnd.android-dir/mms-sms']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'URL Connection to file/http/https/ftp/jar',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'net.URLConnection',
-        'string_or1': 'openConnection',
-        'string_or2': 'connect',
-        'string_or3': 'openStream',
+        'type': StringAndOr.__name__,
+        'match': ['net.URLConnection',
+                  ['openConnection', 'connect', 'openStream']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'JAR URL Connection',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'net.JarURLConnection',
-        'string_or1': 'JarURLConnection',
-        'string_or2': 'jar:',
+        'type': StringAndOr.__name__,
+        'match': ['net.JarURLConnection', ['JarURLConnection', 'jar:']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'HTTPS Connection',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'javax.net.ssl.HttpsURLConnection',
-        'string_or1': 'HttpsURLConnection',
-        'string_or2': 'connect',
+        'type': StringAndOr.__name__,
+        'match': ['javax.net.ssl.HttpsURLConnection',
+                  ['HttpsURLConnection', 'connect']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'URL Connection supports file,http,https,ftp and jar',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'net.URL',
-        'string_or1': 'openConnection',
-        'string_or2': 'openStream',
+        'type': StringAndOr.__name__,
+        'match': ['net.URL', ['openConnection', 'openStream']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Set or Read Clipboard data',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'content.ClipboardManager',
-        'string_or1': 'CLIPBOARD_SERVICE',
-        'string_or2': 'ClipboardManager',
+        'type': StringAndOr.__name__,
+        'match': ['content.ClipboardManager',
+                  ['CLIPBOARD_SERVICE', 'ClipboardManager']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Base64 Encode',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'android.util.Base64',
-        'string_or1': '.encodeToString',
-        'string_or2': '.encode',
+        'type': StringAndOr.__name__,
+        'match': ['android.util.Base64', ['.encodeToString', '.encode']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Base64 Encode',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'android.util.Base64',
-        'string_or1': '.encodeToString',
-        'string_or2': '.encode',
+        'type': StringAndOr.__name__,
+        'match': ['android.util.Base64', ['.encodeToString', '.encode']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Message Digest',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'java.security.MessageDigest',
-        'string_or1': 'MessageDigestSpi',
-        'string_or2': 'MessageDigest',
+        'type': StringAndOr.__name__,
+        'match': ['java.security.MessageDigest',
+                  ['MessageDigestSpi', 'MessageDigest']],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'GPS Location',
-        'type': MatchType.string,
-        'match': Match.string_and_or,
-        'string1': 'android.location',
-        'string_or1': 'getLastKnownLocation(',
-        'string_or2': 'requestLocationUpdates(',
-        'string_or3': 'getLatitude(',
-        'string_or4': 'getLongitude(',
+        'type': StringAndOr.__name__,
+        'match': ['android.location',
+                  ['getLastKnownLocation(', 'requestLocationUpdates(',
+                   'getLatitude(', 'getLongitude(']],
         'input_case': InputCase.exact,
     },
 ]
