+++ StaticAnalyzer/views/android/static_analyzer.py
@@ -42,8 +42,6 @@ from StaticAnalyzer.views.shared_func import (firebase_analysis,
                                               hash_gen, score, unzip,
                                               update_scan_timestamp)
 
-from ..match_command import MatchCommand
-
 from androguard.core.bytecodes import apk
 
 try:
@@ -97,10 +95,6 @@ def static_analyzer(request, api=False):
             app_dic['tools_dir'] = os.path.join(
                 app_dic['dir'], 'StaticAnalyzer/tools/')  # TOOLS DIR
             logger.info('Starting Analysis on : %s', app_dic['app_name'])
-
-            # Will inject the different pattern strategy
-            # when it it will be requested
-            match_command = MatchCommand()
 
             if typ == 'apk':
                 # Check if in DB
@@ -205,8 +199,7 @@ def static_analyzer(request, api=False):
                     code_an_dic = code_analysis(
                         app_dic['app_dir'],
                         man_an_dic['permissons'],
-                        'apk',
-                        match_command)
+                        'apk')
 
                     # Get the strings
                     string_res = strings_jar(
@@ -402,9 +395,7 @@ def static_analyzer(request, api=False):
                         code_an_dic = code_analysis(
                             app_dic['app_dir'],
                             man_an_dic['permissons'],
-                            pro_type,
-                            match_command,
-                        )
+                            pro_type)
                         # Firebase DB Check
                         code_an_dic['firebase'] = firebase_analysis(
                             list(set(code_an_dic['urls_list'])))
