+++ StaticAnalyzer/views/ios/rules/common_rules.py
@@ -38,25 +38,27 @@ Rule Format.
 
 """
 
+from StaticAnalyzer.views.matchers import (
+    SingleRegex,
+    StringAnd,
+    StringOr,
+)
+from StaticAnalyzer.views.rules_properties import (
+    InputCase,
+    Level,
+)
 from StaticAnalyzer.views.standards import (
     CWE,
     OWASP,
     OWASP_MSTG,
 )
-from StaticAnalyzer.views.rules_properties import (
-    InputCase,
-    Level,
-    Match,
-    MatchType,
-)
 
 COMMON_RULES = [
     {
         'desc': 'IP Address disclosure',
-        'type': MatchType.regex,
-        'regex1': r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}',
+        'type': SingleRegex.__name__,
+        'match': r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}',
         'level': Level.warning,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 4.3,
         'cwe': CWE['CWE-200'],
@@ -66,14 +68,13 @@ COMMON_RULES = [
     {
         'desc': ('Files may contain hardcoded sensitive'
                  ' informations like usernames, passwords, keys etc.'),
-        'type': MatchType.regex,
-        'regex1': (r'(password\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
-                   r'(pass\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
-                   r'(username\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
-                   r'(secret\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
-                   r'(key\s*=\s*[\'|\"].+[\'|\"]\s{0,5})'),
+        'type': SingleRegex.__name__,
+        'match': (r'(password\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
+                  r'(pass\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
+                  r'(username\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
+                  r'(secret\s*=\s*[\'|\"].+[\'|\"]\s{0,5})|'
+                  r'(key\s*=\s*[\'|\"].+[\'|\"]\s{0,5})'),
         'level': Level.high,
-        'match': Match.single_regex,
         'input_case': InputCase.lower,
         'cvss': 7.4,
         'cwe': CWE['CWE-312'],
@@ -83,11 +84,9 @@ COMMON_RULES = [
     {
         'desc': ('App uses SQLite Database. '
                  'Sensitive Information should be encrypted.'),
-        'type': MatchType.string,
-        'string1': 'sqlite3_exec',
-        'string2': 'sqlite3_finalize',
+        'type': StringOr.__name__,
+        'match': ['sqlite3_exec', 'sqlite3_finalize'],
         'level': Level.info,
-        'match': Match.string_or,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -97,11 +96,9 @@ COMMON_RULES = [
     {
         'desc': ('User input in "loadHTMLString" '
                  'will result in JavaScript Injection.'),
-        'type': MatchType.string,
-        'string1': 'loadHTMLString',
-        'string2': 'webView',
+        'type': StringAnd.__name__,
+        'match': ['loadHTMLString', 'webView'],
         'level': Level.warning,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 8.8,
         'cwe': CWE['CWE-95'],
@@ -110,58 +107,57 @@ COMMON_RULES = [
     },
     {
         'desc': 'This App may have Jailbreak detection capabilities.',
-        'type': MatchType.string,
-        'string1': '/Applications/Cydia.app',
-        'string2': '/Library/MobileSubstrate/MobileSubstrate.dylib',
-        'string3': '/usr/sbin/sshd',
-        'string4': '/etc/apt',
-        'string5': 'cydia://',
-        'string6': '/var/lib/cydia',
-        'string7': '/Applications/FakeCarrier.app',
-        'string8': '/Applications/Icy.app',
-        'string9': '/Applications/IntelliScreen.app',
-        'string10': '/Applications/SBSettings.app',
-        'string11': ('/Library/MobileSubstrate/DynamicLibraries/'
-                     'LiveClock.plist'),
-        'string12': '/System/Library/LaunchDaemons/com.ikey.bbot.plist',
-        'string13': ('/System/Library/LaunchDaemons/'
-                     'com.saurik.Cydia.Startup.plist'),
-        'string14': '/etc/ssh/sshd_config',
-        'string15': '/private/var/tmp/cydia.log',
-        'string16': '/usr/libexec/ssh-keysign',
-        'string17': '/Applications/MxTube.app',
-        'string18': '/Applications/RockApp.app',
-        'string19': '/Applications/WinterBoard.app',
-        'string20': '/Applications/blackra1n.app',
-        'string21': '/Library/MobileSubstrate/DynamicLibraries/Veency.plist',
-        'string22': '/private/var/lib/apt',
-        'string23': '/private/var/lib/cydia',
-        'string24': '/private/var/mobile/Library/SBSettings/Themes',
-        'string25': '/private/var/stash',
-        'string26': '/usr/bin/sshd',
-        'string27': '/usr/libexec/sftp-server',
-        'string28': '/var/cache/apt',
-        'string29': '/var/lib/apt',
-        'string30': '/usr/sbin/frida-server',
-        'string31': '/usr/bin/cycript',
-        'string32': '/usr/local/bin/cycript',
-        'string33': '/usr/lib/libcycript.dylib',
-        'string34': 'frida-server',
-        'string35': '/etc/apt/sources.list.d/electra.list',
-        'string36': '/etc/apt/sources.list.d/sileo.sources',
-        'string37': '/.bootstrapped_electra',
-        'string38': '/usr/lib/libjailbreak.dylib',
-        'string39': '/jb/lzma',
-        'string40': '/.cydia_no_stash',
-        'string41': '/.installed_unc0ver',
-        'string42': '/jb/offsets.plist',
-        'string43': '/usr/share/jailbreak/injectme.plist',
-        'string44': '/Library/MobileSubstrate/MobileSubstrate.dylib',
-        'string45': '/usr/libexec/cydia/firmware.sh',
-        'string46': '/private/var/cache/apt/',
-        'string47': '/Library/MobileSubstrate/CydiaSubstrate.dylib',
+        'type': StringOr.__name__,
+        'match': ['/Applications/Cydia.app',
+                  '/Library/MobileSubstrate/MobileSubstrate.dylib',
+                  '/usr/sbin/sshd',
+                  '/etc/apt',
+                  'cydia://',
+                  '/var/lib/cydia',
+                  '/Applications/FakeCarrier.app',
+                  '/Applications/Icy.app',
+                  '/Applications/IntelliScreen.app',
+                  '/Applications/SBSettings.app',
+                  ('/Library/MobileSubstrate/DynamicLibraries/'
+                   'LiveClock.plist'),
+                  '/System/Library/LaunchDaemons/com.ikey.bbot.plist',
+                  ('/System/Library/LaunchDaemons/'
+                   'com.saurik.Cydia.Startup.plist'),
+                  '/etc/ssh/sshd_config',
+                  '/private/var/tmp/cydia.log',
+                  '/usr/libexec/ssh-keysign',
+                  '/Applications/MxTube.app',
+                  '/Applications/RockApp.app',
+                  '/Applications/WinterBoard.app',
+                  '/Applications/blackra1n.app',
+                  '/Library/MobileSubstrate/DynamicLibraries/Veency.plist',
+                  '/private/var/lib/apt',
+                  '/private/var/lib/cydia',
+                  '/private/var/mobile/Library/SBSettings/Themes',
+                  '/private/var/stash',
+                  '/usr/bin/sshd',
+                  '/usr/libexec/sftp-server',
+                  '/var/cache/apt',
+                  '/var/lib/apt',
+                  '/usr/sbin/frida-server',
+                  '/usr/bin/cycript',
+                  '/usr/local/bin/cycript',
+                  '/usr/lib/libcycript.dylib',
+                  'frida-server',
+                  '/etc/apt/sources.list.d/electra.list',
+                  '/etc/apt/sources.list.d/sileo.sources',
+                  '/.bootstrapped_electra',
+                  '/usr/lib/libjailbreak.dylib',
+                  '/jb/lzma',
+                  '/.cydia_no_stash',
+                  '/.installed_unc0ver',
+                  '/jb/offsets.plist',
+                  '/usr/share/jailbreak/injectme.plist',
+                  '/Library/MobileSubstrate/MobileSubstrate.dylib',
+                  '/usr/libexec/cydia/firmware.sh',
+                  '/private/var/cache/apt/',
+                  '/Library/MobileSubstrate/CydiaSubstrate.dylib'],
         'level': Level.good,
-        'match': Match.string_or,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
