+++ StaticAnalyzer/views/ios/rules/ios_apis.py
@@ -34,104 +34,92 @@ Rule Format.
 
 from StaticAnalyzer.views.rules_properties import (
     InputCase,
-    Match,
-    MatchType,
+)
+from StaticAnalyzer.views.matchers import (
+    RegexAnd,
+    SingleRegex,
+    SingleString,
+    StringAnd,
 )
 
 CODE_APIS = [
     {
         'desc': 'Network Calls',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'URLSession|CFStream|NSStream',
+        'type': SingleRegex.__name__,
+        'match': r'URLSession|CFStream|NSStream',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Keychain Access.',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': (r'Keychain|SecItemAdd|SecItemUpdate'
-                   r'SecItemCopy| kSecAttr'),
+        'type': SingleRegex.__name__,
+        'match': (r'Keychain|SecItemAdd|SecItemUpdate'
+                  r'SecItemCopy| kSecAttr'),
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView Component',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'UIWebView|WKWebView',
+        'type': SingleRegex.__name__,
+        'match': r'UIWebView|WKWebView',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'SFSafariViewController Component',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'SFSafariViewController',
+        'type': SingleRegex.__name__,
+        'match': r'SFSafariViewController',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Encryption API',
-        'type': MatchType.regex,
-        'match': Match.single_regex,
-        'regex1': r'CommonCrypto|CryptoKit',
+        'type': SingleRegex.__name__,
+        'match': r'CommonCrypto|CryptoKit',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView Load Request',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'loadRequest',
-        'string2': 'webView',
+        'type': StringAnd.__name__,
+        'match': ['loadRequest', 'webView'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'WebView Load HTML String',
-        'type': MatchType.string,
-        'match': Match.string_and,
-        'string1': 'loadHTMLString',
-        'string2': 'webView',
+        'type': StringAnd.__name__,
+        'match': ['loadHTMLString', 'webView'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Cookie Storage',
-        'type': MatchType.regex,
-        'match': Match.regex_and,
-        'regex1': r'HTTPCookieStorage',
-        'regex2': r'shared|sharedHTTPCookieStorage',
+        'type': RegexAnd.__name__,
+        'match': [r'HTTPCookieStorage', r'shared|sharedHTTPCookieStorage'],
         'input_case': InputCase.exact,
     },
     {
         'desc': 'UIPasteboard',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'UIPasteboard',
+        'type': SingleString.__name__,
+        'match': 'UIPasteboard',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'UserDefaults',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'UserDefaults',
+        'type': SingleString.__name__,
+        'match': 'UserDefaults',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'CoreData',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'CoreData',
+        'type': SingleString.__name__,
+        'match': 'CoreData',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'Local Authentication Framework',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'LAContext',
+        'type': SingleString.__name__,
+        'match': 'LAContext',
         'input_case': InputCase.exact,
     },
     {
         'desc': 'UIActivity Sharing',
-        'type': MatchType.string,
-        'match': Match.single_string,
-        'string1': 'UIActivity',
+        'type': SingleString.__name__,
+        'match': 'UIActivity',
         'input_case': InputCase.exact,
     },
 ]
