+++ StaticAnalyzer/views/ios/rules/objc_rules.py
@@ -37,28 +37,31 @@ Rule Format.
    b. regex<no> - regex1, regex2, regex3
 
 """
+from StaticAnalyzer.views.ios.rules import common_rules
+from StaticAnalyzer.views.matchers import (
+    SingleRegex,
+    SingleString,
+    StringAnd,
+    StringAndOr,
+)
+from StaticAnalyzer.views.rules_properties import (
+    InputCase,
+    Level,
+)
 from StaticAnalyzer.views.standards import (
     CWE,
     OWASP,
     OWASP_MSTG,
-)
-from StaticAnalyzer.views.ios.rules import common_rules
-from StaticAnalyzer.views.rules_properties import (
-    InputCase,
-    Level,
-    Match,
-    MatchType,
 )
 
 OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('The App may contain banned API(s). '
                  'These API(s) are insecure and must not be used.'),
-        'type': MatchType.regex,
-        'regex1': (r'strcpy|memcpy|strcat|strncat|'
-                   r'strncpy|sprintf|vsprintf|gets'),
+        'type': SingleRegex.__name__,
+        'match': (r'strcpy|memcpy|strcat|strncat|'
+                  r'strncpy|sprintf|vsprintf|gets'),
         'level': Level.high,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 2.2,
         'cwe': CWE['CWE-676'],
@@ -68,16 +71,15 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('App allows self signed or invalid '
                  'SSL certificates. App is vulnerable to MITM attacks.'),
-        'type': MatchType.regex,
-        'regex1': (r'canAuthenticateAgainstProtectionSpace|'
-                   r'continueWithoutCredentialForAuthenticationChallenge|'
-                   r'kCFStreamSSLAllowsExpiredCertificates|'
-                   r'kCFStreamSSLAllowsAnyRoot|'
-                   r'kCFStreamSSLAllowsExpiredRoots|'
-                   r'validatesSecureCertificate\s*=\s*(no|NO)|'
-                   r'allowInvalidCertificates\s*=\s*(YES|yes)'),
+        'type': SingleRegex.__name__,
+        'match': (r'canAuthenticateAgainstProtectionSpace|'
+                  r'continueWithoutCredentialForAuthenticationChallenge|'
+                  r'kCFStreamSSLAllowsExpiredCertificates|'
+                  r'kCFStreamSSLAllowsAnyRoot|'
+                  r'kCFStreamSSLAllowsExpiredRoots|'
+                  r'validatesSecureCertificate\s*=\s*(no|NO)|'
+                  r'allowInvalidCertificates\s*=\s*(YES|yes)'),
         'level': Level.high,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 7.4,
         'cwe': CWE['CWE-295'],
@@ -87,12 +89,11 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('UIWebView in App ignore SSL errors and accept'
                  ' any SSL Certificate. App is vulnerable to MITM attacks.'),
-        'type': MatchType.regex,
-        'regex1': (r'setAllowsAnyHTTPSCertificate:\s*YES|'
-                   r'allowsAnyHTTPSCertificateForHost|'
-                   r'loadingUnvalidatedHTTPSPage\s*=\s*(YES|yes)'),
+        'type': SingleRegex.__name__,
+        'match': (r'setAllowsAnyHTTPSCertificate:\s*YES|'
+                  r'allowsAnyHTTPSCertificateForHost|'
+                  r'loadingUnvalidatedHTTPSPage\s*=\s*(YES|yes)'),
         'level': Level.high,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 7.4,
         'cwe': CWE['CWE-295'],
@@ -102,10 +103,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('The App logs information. '
                  'Sensitive information should never be logged.'),
-        'type': MatchType.regex,
-        'regex1': r'NSLog|NSAssert|fprintf|fprintf|Logging',
+        'type': SingleRegex.__name__,
+        'match': r'NSLog|NSAssert|fprintf|fprintf|Logging',
         'level': Level.info,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 7.5,
         'cwe': CWE['CWE-532'],
@@ -115,11 +115,10 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('This app listens to Clipboard changes. '
                  'Some malwares also listen to Clipboard changes.'),
-        'type': MatchType.regex,
-        'regex1': (r'UIPasteboardChangedNotification|'
-                   r'generalPasteboard\]\.string'),
+        'type': SingleRegex.__name__,
+        'match': (r'UIPasteboardChangedNotification|'
+                  r'generalPasteboard\]\.string'),
         'level': Level.warning,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -129,10 +128,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('Untrusted user input to "NSTemporaryDirectory()"'
                  ' will result in path traversal vulnerability.'),
-        'type': MatchType.string,
-        'string1': 'NSTemporaryDirectory(),',
+        'type': SingleString.__name__,
+        'match': 'NSTemporaryDirectory(),',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 7.5,
         'cwe': CWE['CWE-22'],
@@ -143,10 +141,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
         'desc': ('File is stored in an encrypted format on '
                  'disk and cannot be read from or written to '
                  'while the device is locked or booting.'),
-        'type': MatchType.string,
-        'string1': 'NSFileProtectionComplete',
+        'type': SingleString.__name__,
+        'match': 'NSFileProtectionComplete',
         'level': Level.good,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -156,10 +153,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('File is stored in an encrypted format '
                  'on disk after it is closed.'),
-        'type': MatchType.string,
-        'string1': 'NSFileProtectionCompleteUnlessOpen',
+        'type': SingleString.__name__,
+        'match': 'NSFileProtectionCompleteUnlessOpen',
         'level': Level.good,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -170,12 +166,11 @@ OBJC_RULES = common_rules.COMMON_RULES + [
         'desc': ('File is stored in an encrypted format '
                  'on disk and cannot be accessed until after '
                  'the device has booted.'),
-        'type': MatchType.string,
-        'string1': (
+        'type': SingleString.__name__,
+        'match': (
             'NSFileProtectionComplete'
             'UntilFirstUserAuthentication'),
         'level': Level.good,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -185,10 +180,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('The file has no special protections '
                  'associated with it.'),
-        'type': MatchType.string,
-        'string1': 'NSFileProtectionNone',
+        'type': SingleString.__name__,
+        'match': 'NSFileProtectionNone',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 4.3,
         'cwe': CWE['CWE-311'],
@@ -197,12 +191,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'SFAntiPiracy Jailbreak checks found',
-        'type': MatchType.string,
-        'string1': 'SFAntiPiracy.h',
-        'string2': 'SFAntiPiracy',
-        'string3': 'isJailbroken',
+        'type': StringAnd.__name__,
+        'match': ['SFAntiPiracy.h', 'SFAntiPiracy', 'isJailbroken'],
         'level': Level.good,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -211,12 +202,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'SFAntiPiracy Piracy checks found',
-        'type': MatchType.string,
-        'string1': 'SFAntiPiracy.h',
-        'string2': 'SFAntiPiracy',
-        'string3': 'isPirated',
+        'type': StringAnd.__name__,
+        'match': ['SFAntiPiracy.h', 'SFAntiPiracy', 'isPirated'],
         'level': Level.good,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -225,11 +213,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'MD5 is a weak hash known to have hash collisions.',
-        'type': MatchType.string,
-        'string1': 'CommonDigest.h',
-        'string2': 'CC_MD5',
+        'type': StringAnd.__name__,
+        'match': ['CommonDigest.h', 'CC_MD5'],
         'level': Level.high,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 7.4,
         'cwe': CWE['CWE-327'],
@@ -238,11 +224,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'SHA1 is a weak hash known to have hash collisions.',
-        'type': MatchType.string,
-        'string1': 'CommonDigest.h',
-        'string2': 'CC_SHA1',
+        'type': StringAnd.__name__,
+        'match': ['CommonDigest.h', 'CC_SHA1'],
         'level': Level.high,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -253,11 +237,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
         'desc': ('The App uses ECB mode in Cryptographic encryption algorithm.'
                  ' ECB mode is known to be weak as it results in the same'
                  ' ciphertext for identical blocks of plaintext.'),
-        'type': MatchType.string,
-        'string1': 'kCCOptionECBMode',
-        'string2': 'kCCAlgorithmAES',
+        'type': StringAnd.__name__,
+        'match': ['kCCOptionECBMode', 'kCCAlgorithmAES'],
         'level': Level.high,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -266,11 +248,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'The App has anti-debugger code using ptrace() ',
-        'type': MatchType.string,
-        'string1': 'ptrace_ptr',
-        'string2': 'PT_DENY_ATTACH',
+        'type': StringAnd.__name__,
+        'match': ['ptrace_ptr', 'PT_DENY_ATTACH'],
         'level': Level.info,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -279,12 +259,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'This App has anti-debugger code using Mach Exception Ports.',
-        'type': MatchType.string,
-        'string1': 'mach/mach_init.h',
-        'string_or1': 'MACH_PORT_VALID',
-        'string_or2': 'mach_task_self()',
+        'type': StringAndOr.__name__,
+        'match': ['mach/mach_init.h', ['MACH_PORT_VALID', 'mach_task_self()']],
         'level': Level.info,
-        'match': Match.string_and_or,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -295,12 +272,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
         'desc': ('This App copies data to clipboard. Sensitive data should'
                  ' not be copied to clipboard as other applications'
                  ' can access it.'),
-        'type': MatchType.string,
-        'string1': 'UITextField',
-        'string_or1': '@select(cut:)',
-        'string_or2': '@select(copy:)',
+        'type': StringAndOr.__name__,
+        'match': ['UITextField', ['@select(cut:)', '@select(copy:)']],
         'level': Level.info,
-        'match': Match.string_and_or,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
@@ -310,10 +284,9 @@ OBJC_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('App uses Realm Database. '
                  'Sensitive Information should be encrypted.'),
-        'type': MatchType.string,
-        'string1': '[realm transactionWithBlock:',
+        'type': SingleString.__name__,
+        'match': '[realm transactionWithBlock:',
         'level': Level.info,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': '',
