+++ StaticAnalyzer/views/ios/rules/swift_rules.py
@@ -37,27 +37,34 @@ Rule Format.
    b. regex<no> - regex1, regex2, regex3
 
 """
+from StaticAnalyzer.views.ios.rules import common_rules
+from StaticAnalyzer.views.matchers import (
+    RegexAnd,
+    RegexOr,
+    SingleRegex,
+    SingleString,
+    StringAnd,
+    StringAndNot,
+    StringAndOr,
+    StringOr,
+)
+from StaticAnalyzer.views.rules_properties import (
+    InputCase,
+    Level,
+)
 from StaticAnalyzer.views.standards import (
     CWE,
     OWASP,
     OWASP_MSTG,
 )
-from StaticAnalyzer.views.rules_properties import (
-    InputCase,
-    Level,
-    Match,
-    MatchType,
-)
-from StaticAnalyzer.views.ios.rules import common_rules
 
 SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('The App logs information. '
                  'Sensitive information should never be logged.'),
-        'type': MatchType.regex,
-        'regex1': r'(print|NSLog|os_log|OSLog|os_signpost)\(.*\)',
+        'type': SingleRegex.__name__,
+        'match': r'(print|NSLog|os_log|OSLog|os_signpost)\(.*\)',
         'level': Level.info,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 7.5,
         'cwe': CWE['CWE-532'],
@@ -66,11 +73,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'SHA1 is a weak hash known to have hash collisions.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)SHA1\(',
-        'regex2': r'CC_SHA1\(',
+        'type': RegexOr.__name__,
+        'match': [r'(?i)SHA1\(', r'CC_SHA1\('],
         'level': Level.high,
-        'match': Match.regex_or,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -79,11 +84,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'MD2 is a weak hash known to have hash collisions.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)MD2\(',
-        'regex2': r'CC_MD2\(',
+        'type': RegexOr.__name__,
+        'match': [r'(?i)MD2\(', r'CC_MD2\('],
         'level': Level.high,
-        'match': Match.regex_or,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -92,11 +95,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'MD4 is a weak hash known to have hash collisions.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)MD4\(',
-        'regex2': r'CC_MD4\(',
+        'type': RegexOr.__name__,
+        'match': [r'(?i)MD4\(', r'CC_MD4\('],
         'level': Level.high,
-        'match': Match.regex_or,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -105,11 +106,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'MD5 is a weak hash known to have hash collisions.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)MD5\(',
-        'regex2': r'CC_MD5\(',
+        'type': RegexOr.__name__,
+        'match': [r'(?i)MD5\(', r'CC_MD5\('],
         'level': Level.high,
-        'match': Match.regex_or,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -118,11 +117,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'MD6 is a weak hash known to have hash collisions.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)MD6\(',
-        'regex2': r'CC_MD6\(',
+        'type': RegexOr.__name__,
+        'match': [r'(?i)MD6\(', r'CC_MD6\('],
         'level': Level.high,
-        'match': Match.regex_or,
         'input_case': InputCase.exact,
         'cvss': 5.9,
         'cwe': CWE['CWE-327'],
@@ -131,10 +128,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'This App may have custom keyboards disabled.',
-        'type': MatchType.regex,
-        'regex1': r'extensionPointIdentifier == .*\.keyboard',
+        'type': SingleRegex.__name__,
+        'match': r'extensionPointIdentifier == .*\.keyboard',
         'level': Level.good,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -144,10 +140,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('Keyboard cache should be disabled for all '
                  'sensitve data inputs.'),
-        'type': MatchType.regex,
-        'regex1': r'.autocorrectionType = .no',
+        'type': SingleRegex.__name__,
+        'match': r'.autocorrectionType = .no',
         'level': Level.good,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -157,10 +152,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('It is recommended to use WKWebView instead '
                  'of SFSafariViewController or UIViewView.'),
-        'type': MatchType.regex,
-        'regex1': r'UIViewView|SFSafariViewController',
+        'type': SingleRegex.__name__,
+        'match': r'UIViewView|SFSafariViewController',
         'level': Level.warning,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -169,13 +163,12 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'This App may have Reverse enginering detection capabilities.',
-        'type': MatchType.string,
-        'string1': '\"FridaGadget\"',
-        'string2': '\"cynject\"',
-        'string3': '\"libcycript\"',
-        'string4': '\"/usr/sbin/frida-server\"',
+        'type': StringAnd.__name__,
+        'match': ['\"FridaGadget\"',
+                  '\"cynject\"',
+                  '\"libcycript\"',
+                  '\"/usr/sbin/frida-server\"'],
         'level': Level.good,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -184,10 +177,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'This App may have Emulator detection capabilities.',
-        'type': MatchType.string,
-        'string1': '\"SIMULATOR_DEVICE_NAME\"',
+        'type': SingleString.__name__,
+        'match': '\"SIMULATOR_DEVICE_NAME\"',
         'level': Level.good,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -197,10 +189,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('Biometric authentication should be based '
                  'on Keychain, not based on bool.'),
-        'type': MatchType.regex,
-        'regex1': r'\.evaluatePolicy\(.deviceOwnerAuthentication',
+        'type': SingleRegex.__name__,
+        'match': r'\.evaluatePolicy\(.deviceOwnerAuthentication',
         'level': Level.warning,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -209,10 +200,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'The file has no special protections associated with it.',
-        'type': MatchType.regex,
-        'regex1': r'(?i)\.noFileProtection',
+        'type': SingleRegex.__name__,
+        'match': r'(?i)\.noFileProtection',
         'level': Level.high,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 4.3,
         'cwe': CWE['CWE-311'],
@@ -222,10 +212,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('This application uses UIPasteboard, improper use '
                  'of this class can lead to security issues.'),
-        'type': MatchType.string,
-        'string1': 'UIPasteboard',
+        'type': SingleString.__name__,
+        'match': 'UIPasteboard',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -234,10 +223,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'Usage of generalPasteboard should be avoided.',
-        'type': MatchType.string,
-        'string1': 'UIPasteboard.generalPasteboard',
+        'type': SingleString.__name__,
+        'match': 'UIPasteboard.generalPasteboard',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -248,10 +236,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
         'desc': ('The app should not export sensitive functionality via '
                  'custom URL schemes, unless these '
                  'mechanisms are properly protected.'),
-        'type': MatchType.string,
-        'string1': 'CFBundleURLSchemes',
+        'type': SingleString.__name__,
+        'match': 'CFBundleURLSchemes',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -260,10 +247,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'Some UI controls have secure text entry configured.',
-        'type': MatchType.string,
-        'string1': '.secureTextEntry = true',
+        'type': SingleString.__name__,
+        'match': '.secureTextEntry = true',
         'level': Level.good,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -272,11 +258,10 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'This App may have Certificate Pinning mechanism.',
-        'type': MatchType.string,
-        'string1': 'PinnedCertificatesTrustEvaluator',
-        'string2': 'TrustKit.initSharedInstance',
+        'type': StringOr.__name__,
+        'match': ['PinnedCertificatesTrustEvaluator',
+                  'TrustKit.initSharedInstance'],
         'level': Level.good,
-        'match': Match.string_or,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': CWE['CWE-295'],
@@ -286,10 +271,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('App uses Realm Database. '
                  'Sensitive Information should be encrypted.'),
-        'type': MatchType.string,
-        'string1': 'realm.write',
+        'type': SingleString.__name__,
+        'match': 'realm.write',
         'level': Level.info,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0,
         'cwe': CWE['CWE-311'],
@@ -299,11 +283,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('App uses CoreData Database. '
                  'Sensitive Information should be encrypted.'),
-        'type': MatchType.string,
-        'string1': 'NSManagedObjectContext',
-        'string2': '.save()',
+        'type': StringAnd.__name__,
+        'match': ['NSManagedObjectContext', '.save()'],
         'level': Level.info,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': CWE['CWE-311'],
@@ -312,11 +294,10 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'Used Realm database has configured encryption.',
-        'type': MatchType.string,
-        'string1': 'Realm.Configuration(encryptionKey:',
-        'string2': 'Realm(configuration:',
+        'type': StringAnd.__name__,
+        'match': ['Realm.Configuration(encryptionKey:',
+                  'Realm(configuration:'],
         'level': Level.good,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': CWE['CWE-311'],
@@ -325,11 +306,10 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'Copy feature may be disabled in some UI controls.',
-        'type': MatchType.regex,
-        'regex1': r'canPerformAction',
-        'regex2': r'UIResponderStandardEditActions\.copy|\.copy',
+        'type': RegexAnd.__name__,
+        'match': [r'canPerformAction',
+                  r'UIResponderStandardEditActions\.copy|\.copy'],
         'level': Level.info,
-        'match': Match.regex_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -338,10 +318,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     },
     {
         'desc': 'Opening App with URLs can lead to potencial security leaks.',
-        'type': MatchType.regex,
-        'regex1': r'func application\(.*open url: URL',
+        'type': SingleRegex.__name__,
+        'match': r'func application\(.*open url: URL',
         'level': Level.warning,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': CWE['CWE-939'],
@@ -351,10 +330,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('Sensitive data shouldn\'t be included in backups generated '
                  'by the mobile operating system.'),
-        'type': MatchType.regex,
-        'regex1': r'kSecAttrAccessible[a-zA-Z]*ThisDeviceOnly',
+        'type': SingleRegex.__name__,
+        'match': r'kSecAttrAccessible[a-zA-Z]*ThisDeviceOnly',
         'level': Level.good,
-        'match': Match.single_regex,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -364,11 +342,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
     {
         'desc': ('JavaScript should be disabled in WebViews unless '
                  'explicitly required.'),
-        'type': MatchType.string,
-        'string1': 'WKWebView',
-        'string2': '.javaScriptEnabled = false',
+        'type': StringAndNot.__name__,
+        'match': ['WKWebView', '.javaScriptEnabled = false'],
         'level': Level.warning,
-        'match': Match.string_and_not,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -376,14 +352,12 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
         'owasp-mstg': OWASP_MSTG['platform-5'],
     },
     {
-        'desc': ('TLS 1.3 should be used. '
-                 'Detected old version'),
-        'type': MatchType.string,
-        'string1': '.TLSMinimumSupportedProtocolVersion',
-        'string_or1': 'tls_protocol_version_t.TLSv10',
-        'string_or2': 'tls_protocol_version_t.TLSv11',
+        'desc': 'TLS 1.3 should be used. Detected old version',
+        'type': StringAndOr.__name__,
+        'match': ['.TLSMinimumSupportedProtocolVersion',
+                  ['tls_protocol_version_t.TLSv10',
+                   'tls_protocol_version_t.TLSv11']],
         'level': Level.high,
-        'match': Match.string_and_or,
         'input_case': InputCase.exact,
         'cvss': 7.5,
         'cwe': CWE['CWE-757'],
@@ -391,13 +365,11 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
         'owasp-mstg': OWASP_MSTG['network-2'],
     },
     {
-        'desc': ('TLS 1.3 should be used. '
-                 'Detected old version - TLS 1.2.'),
-        'type': MatchType.string,
-        'string1': '.TLSMinimumSupportedProtocolVersion',
-        'string2': 'tls_protocol_version_t.TLSv12',
+        'desc': 'TLS 1.3 should be used. Detected old version - TLS 1.2.',
+        'type': StringAnd.__name__,
+        'match': ['.TLSMinimumSupportedProtocolVersion',
+                  'tls_protocol_version_t.TLSv12'],
         'level': Level.warning,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -405,13 +377,11 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
         'owasp-mstg': OWASP_MSTG['network-2'],
     },
     {
-        'desc': ('DTLS 1.2 should be used. '
-                 'Detected old version - DTLS 1.0.'),
-        'type': MatchType.string,
-        'string1': '.TLSMinimumSupportedProtocolVersion',
-        'string2': 'tls_protocol_version_t.DTLSv10',
+        'desc': 'DTLS 1.2 should be used. Detected old version - DTLS 1.0.',
+        'type': StringAnd.__name__,
+        'match': ['.TLSMinimumSupportedProtocolVersion',
+                  'tls_protocol_version_t.DTLSv10'],
         'level': Level.warning,
-        'match': Match.string_and,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
@@ -422,10 +392,9 @@ SWIFT_RULES = common_rules.COMMON_RULES + [
         'desc': ('Use of deprecated property tlsMinimumSupportedProtocol. '
                  'To avoid potential security risks, use '
                  'tlsMinimumSupportedProtocolVersion'),
-        'type': MatchType.string,
-        'string1': '.tlsMinimumSupportedProtocol',
+        'type': SingleString.__name__,
+        'match': '.tlsMinimumSupportedProtocol',
         'level': Level.warning,
-        'match': Match.single_string,
         'input_case': InputCase.exact,
         'cvss': 0.0,
         'cwe': '',
