+++ StaticAnalyzer/views/ios/static_analyzer.py
@@ -35,8 +35,6 @@ from StaticAnalyzer.views.shared_func import (
     update_scan_timestamp,
 )
 
-from ..match_command import MatchCommand
-
 logger = logging.getLogger(__name__)
 
 ##############################################################
@@ -72,10 +70,6 @@ def static_analyzer_ios(request, api=False):
                 settings.UPLD_DIR, app_dict['md5_hash'] + '/')  # APP DIRECTORY
             tools_dir = os.path.join(
                 app_dict['directory'], 'StaticAnalyzer/tools/ios/')
-
-            # Will inject the different pattern strategy
-            # when it it will be requested
-            match_command = MatchCommand()
 
             if file_type == 'ipa':
                 # DB
@@ -193,8 +187,7 @@ def static_analyzer_ios(request, api=False):
                     infoplist_dict = plist_analysis(app_dict['app_dir'], True)
                     app_dict['appstore'] = app_search(infoplist_dict.get('id'))
                     code_analysis_dic = ios_source_analysis(
-                        app_dict['app_dir'],
-                        match_command)
+                        app_dict['app_dir'])
                     # Get App Icon
                     app_dict['icon_found'] = get_icon_source(
                         app_dict['md5_hash'],
