+++ StaticAnalyzer/views/matchers.py
@@ -5,6 +5,19 @@ import logging
 logger = logging.getLogger(__name__)
 
 
+class MatchCommand:
+
+    def __init__(self):
+        self.patterns = {}
+
+    def find_match(self, pattern_name, content, rule, perms):
+        if pattern_name not in self.patterns:
+            logger.debug('adding %s in the pool of patterns', pattern_name)
+            pattern_class = globals()[pattern_name]
+            self.patterns[pattern_name] = pattern_class()
+        return self.patterns[pattern_name].perform_search(content, rule, perms)
+
+
 class MatchStrategy(ABC):
 
     @abstractclassmethod
@@ -30,7 +43,7 @@ class RegexAnd(MatchStrategy):
             logger.debug('wrong regex type, switching to single regex')
             return SingleRegex().perform_search(content, rule, perms)
         for regex in rule['match']:
-            if bool(re.findall(regex, content)) is False:
+            if not bool(re.findall(regex, content)):
                 found = False
                 break
         return found
